// src/generated/clayers/clayersResponses.ts
/**
 * Generated by @openapi-codegen
 *
 * @version 1.0
 */
export type Layer = {
  data: {
    type: string;
    id: string;
    attributes: {
      isPublic: boolean;
      name: string;
      url: string;
      enabled: boolean;
      owner: {
        type: "USER" | "TEAM";
      };
      /**
       * @format date-time
       */
      createdAt: string;
    };
  };
};

export type Layers = {
  /**
   * @uniqueItems true
   */
  data: {
    type: string;
    id: string;
    attributes?: {
      isPublic: boolean;
      name: string;
      url: string;
      enabled: boolean;
      owner: {
        type: "USER" | "TEAM";
      };
      /**
       * @format date-time
       */
      createdAt: string;
    };
  }[];
};

export type Error = {
  /**
   * @uniqueItems true
   * @minItems 1
   */
  errors: {
    status: number;
    /**
     * @minLength 1
     */
    detail: string;
  }[];
};

export type Healthcheck = {
  uptime?: number;
};